Fix bugs when discarding bonus cards
There was bugs both in the server and in the client when discarding
bonus cards. Firstly, the server sometimes crashed when discarding a
card, this is now hopefully fixed. Secondly, the client sometimes kept
the discarded cards in the interface. This is now fixed along a
slight card-oriented refactorisation of the client.