Create a new Response for new turns
Before, the UpdateBoard Response was used both to synchronize clients
and the server and to tell clients the turn is over and to close all
popups. The problem is, UpdateBoard isn't necessarily sent when there is
a new turn, it is also sent when a card is played. As such, I created a
new Response NewTurn that will tell the client to close all popups, and
UpdateBoard will now only synchronize clients and the server.